+++ libraries/provider_database_mysql.rb
@@ -66,6 +66,7 @@ class Chef
               db.select_db(@new_resource.database_name) if @new_resource.database_name
               Chef::Log.debug("#{@new_resource}: Performing query [#{new_resource.sql_query}]")
               db.query(@new_resource.sql_query)
+              db.next_result while db.next_result
               @new_resource.updated_by_last_action(true)
             ensure
               close
@@ -74,6 +75,7 @@ class Chef
         end
 
         private
+        
         def exists?
           db.list_dbs.include?(@new_resource.database_name)
         end
@@ -97,7 +99,6 @@ class Chef
           @db.close rescue nil
           @db = nil
         end
-
       end
     end
   end