+++ libraries/provider_database_postgresql_user.rb
@@ -1,6 +1,7 @@
 #
 # Author:: Seth Chisamore (<schisamo@opscode.com>)
 # Author:: Lamont Granquist (<lamont@opscode.com>)
+# Author:: Marco Betti (<m.betti@gmail.com>)
 # Copyright:: Copyright (c) 2011 Opscode, Inc.
 # License:: Apache License, Version 2.0
 #
@@ -36,7 +37,9 @@ class Chef
         def action_create
           unless exists?
             begin
-              db("template1").query("CREATE USER \"#{@new_resource.username}\" WITH PASSWORD '#{@new_resource.password}'")
+              statement = "CREATE USER \"#{@new_resource.username}\""
+              statement += " WITH PASSWORD '#{@new_resource.password}'" if @new_resource.password
+              db('template1').query(statement)
               @new_resource.updated_by_last_action(true)
             ensure
               close
@@ -47,7 +50,7 @@ class Chef
         def action_drop
           if exists?
             begin
-              db("template1").query("DROP USER \"#{@new_resource.username}\"")
+              db('template1').query("DROP USER \"#{@new_resource.username}\"")
               @new_resource.updated_by_last_action(true)
             ensure
               close
@@ -67,16 +70,26 @@ class Chef
           end
         end
 
+        def action_grant_schema
+          begin
+            grant_statement = "GRANT #{@new_resource.privileges.join(', ')} ON SCHEMA \"#{@new_resource.schema_name}\" TO \"#{@new_resource.username}\""
+            Chef::Log.info("#{@new_resource}: granting access with statement [#{grant_statement}]")
+            db(@new_resource.database_name).query(grant_statement)
+            @new_resource.updated_by_last_action(true)
+          ensure
+            close
+          end
+        end
+
         private
         def exists?
           begin
-            exists = db("template1").query("SELECT * FROM pg_user WHERE usename='#{@new_resource.username}'").num_tuples != 0
+            exists = db('template1').query("SELECT * FROM pg_user WHERE usename='#{@new_resource.username}'").num_tuples != 0
           ensure
             close
           end
           exists
         end
-
       end
     end
   end