+++ libraries/resource_database_user.rb
@@ -21,8 +21,7 @@ require File.join(File.dirname(__FILE__), 'resource_database')
 class Chef
   class Resource
     class DatabaseUser < Chef::Resource::Database
-
-      def initialize(name, run_context=nil)
+      def initialize(name, run_context = nil)
         super
         @resource_name = :database_user
         @username = name
@@ -37,7 +36,7 @@ class Chef
         @action = :create
       end
 
-      def database_name(arg=nil)
+      def database_name(arg = nil)
         set_or_return(
           :database_name,
           arg,
@@ -45,7 +44,7 @@ class Chef
         )
       end
 
-      def username(arg=nil)
+      def username(arg = nil)
         set_or_return(
           :username,
           arg,
@@ -53,16 +52,15 @@ class Chef
         )
       end
 
-      def password(arg=nil)
+      def password(arg = nil)
         set_or_return(
           :password,
           arg,
-          :kind_of => String,
-          :required => true
+          :kind_of => String
         )
       end
 
-      def table(arg=nil)
+      def table(arg = nil)
         set_or_return(
           :table,
           arg,
@@ -70,7 +68,7 @@ class Chef
         )
       end
 
-      def host(arg=nil)
+      def host(arg = nil)
         set_or_return(
           :host,
           arg,
@@ -78,7 +76,7 @@ class Chef
         )
       end
 
-      def privileges(arg=nil)
+      def privileges(arg = nil)
         set_or_return(
           :privileges,
           arg,
@@ -86,14 +84,13 @@ class Chef
         )
       end
 
-      def grant_option(arg=nil)
+      def grant_option(arg = nil)
         set_or_return(
           :grant_option,
           arg,
-          :kind_of => [ TrueClass, FalseClass ], :default => false
+          :kind_of => [TrueClass, FalseClass], :default => false
         )
       end
-
     end
   end
 end