+++ libraries/resource_mysql_database_user.rb
@@ -22,13 +22,11 @@ require File.join(File.dirname(__FILE__), 'provider_database_mysql_user')
 class Chef
   class Resource
     class MysqlDatabaseUser < Chef::Resource::DatabaseUser
-
-      def initialize(name, run_context=nil)
+      def initialize(name, run_context = nil)
         super
         @resource_name = :mysql_database_user
         @provider = Chef::Provider::Database::MysqlUser
       end
-
     end
   end
 end