+++ libraries/resource_postgresql_database_schema.rb
@@ -1,6 +1,6 @@
 #
-# Author:: Seth Chisamore (<schisamo@opscode.com>)
-# Copyright:: Copyright (c) 2011 Opscode, Inc.
+# Author:: Marco Betti (<m.betti@gmail.com>)
+# Copyright:: Copyright (c) 2013 Opscode, Inc.
 # License:: Apache License, Version 2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,15 +17,26 @@
 #
 
 require File.join(File.dirname(__FILE__), 'resource_database')
-require File.join(File.dirname(__FILE__), 'provider_database_sql_server')
+require File.join(File.dirname(__FILE__), 'provider_database_postgresql_schema')
 
 class Chef
   class Resource
-    class SqlServerDatabase < Chef::Resource::Database
+    class PostgresqlDatabaseSchema < Chef::Resource::Database
       def initialize(name, run_context = nil)
         super
-        @resource_name = :sql_server_database
-        @provider = Chef::Provider::Database::SqlServer
+        @resource_name = :postgresql_database_schema
+        @schema_name = name
+        @allowed_actions.push(:create, :drop)
+        @action = :create
+        @provider = Chef::Provider::Database::PostgresqlSchema
+      end
+
+      def schema_name(arg = nil)
+        set_or_return(
+          :schema_name,
+          arg,
+          :kind_of => String
+        )
       end
     end
   end