+++ recipes/snapshot.rb
@@ -17,22 +17,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-include_recipe "aws"
-include_recipe "xfs"
+include_recipe 'aws'
+include_recipe 'xfs'
 
-%w{ebs_vol_dev db_role app_environment username password aws_access_key_id aws_secret_access_key snapshots_to_keep volume_id}.collect do |key|
-  Chef::Application.fatal!("Required db_snapshot configuration #{key} not found.", -47) unless node.db_snapshot.has_key? key
+%w(ebs_vol_dev db_role app_environment username password aws_access_key_id aws_secret_access_key snapshots_to_keep volume_id).map do |key|
+  Chef::Application.fatal!("Required db_snapshot configuration #{key} not found.", -47) unless node.db_snapshot.key? key
 end
 
-connection_info = {:host => "localhost", :username => node.db_snapshot.username, :password => node.db_snapshot.password}
+connection_info = { :host => 'localhost', :username => node.db_snapshot.username, :password => node.db_snapshot.password }
 
 mysql_database "locking tables for #{node.db_snapshot.app_environment}" do
   connection connection_info
-  sql "flush tables with read lock"
+  sql 'flush tables with read lock'
   action :query
 end
 
-execute "xfs freeze" do
+execute 'xfs freeze' do
   command "xfs_freeze -f #{node.db_snapshot.ebs_vol_dev}"
 end
 
@@ -47,13 +47,13 @@ aws_ebs_volume "#{node.db_snapshot.db_role.first}_#{node.db_snapshot.app_environ
   ignore_failure true # if this fails, continue to unfreeze and unlock
 end
 
-execute "xfs unfreeze" do
+execute 'xfs unfreeze' do
   command "xfs_freeze -u #{node.db_snapshot.ebs_vol_dev}"
 end
 
 mysql_database "unflushing tables for #{node.db_snapshot.app_environment}" do
   connection connection_info
-  sql "unlock tables"
+  sql 'unlock tables'
   action :query
 end